+++ bookmarks/bookmarks/apps.py
@@ -3,5 +3,5 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 
-class CoreConfig(AppConfig):
-    name = 'core'
+class BookmarksConfig(AppConfig):
+    name = 'bookmarks'
