+++ bookmarks/bookmarks/templates/bookmarks/bookmark_list.html
@@ -43,7 +43,7 @@
 
   <div class="row">
     <div class="col-xs-12">
-      {% include 'core/pagination-reverse.html' %}
+      {% include 'bookmarks/pagination-reverse.html' %}
     </div>
   </div>
