+++ bookmarks/bookmarks/templates/bookmarks/charts.html
