+++ bookmarks/config/settings/base.py
@@ -114,7 +114,7 @@ THIRD_PARTY_APPS = (
 
 
 LOCAL_APPS = (
-    'core',
+    'bookmarks',
 )
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
